Fix instant close bug
